// src/common/types/subscription.types.ts
export enum ArewaSubscriptionPlanType {
  AREWA_MONTHLY = "Arewa Monthly Plan",
  AREWA_YEARLY = "Arewa Yearly Plan",
}

export enum SubscriptionPlanIntervalType {
  MONTHLY = "monthly",
  WEEKLY = "weekly",
  DAILY = "daily",
  HOURLY = "hourly",
  QUARTERLY = "quarterly",
  BIANNUALLY = "biannually",
  ANNUALLY = "annually",
}